+++ frontend/app/src/background.ts
@@ -1,19 +1,22 @@
 import path from 'path';
 import {
   app,
-  protocol,
   BrowserWindow,
-  Menu,
-  ipcMain,
-  shell,
   dialog,
+  ipcMain,
+  Menu,
+  MenuItem,
   MenuItemConstructorOptions,
-  MenuItem
+  protocol,
+  shell
 } from 'electron';
 import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer';
 import windowStateKeeper from 'electron-window-state';
 import { createProtocol } from 'vue-cli-plugin-electron-builder/lib';
+import { startHttp, stopHttp } from '@/electron-main/http';
+import { selectPort } from '@/electron-main/port-utils';
 import PyHandler from './py-handler';
+import Timeout = NodeJS.Timeout;
 
 const isDevelopment = process.env.NODE_ENV !== 'production';
 const isMac = process.platform === 'darwin';
@@ -329,6 +332,25 @@ app.on('ready', async () => {
   ipcMain.on('OPEN_DIRECTORY', async (event, args) => {
     const directory = await select(args, 'openDirectory');
     event.sender.send('OPEN_DIRECTORY', directory);
+  });
+  let importTimeout: Timeout;
+  ipcMain.on('METAMASK_IMPORT', async (event, _args) => {
+    try {
+      const port = startHttp(
+        addresses => event.sender.send('METAMASK_IMPORT', { addresses }),
+        await selectPort(40000)
+      );
+      await shell.openExternal(`http://localhost:${port}`);
+      if (importTimeout) {
+        clearTimeout(importTimeout);
+      }
+      importTimeout = setTimeout(() => {
+        stopHttp();
+        event.sender.send('METAMASK_IMPORT', { error: 'waiting timeout' });
+      }, 120000);
+    } catch (e) {
+      event.sender.send('METAMASK_IMPORT', { error: e.message });
+    }
   });
   createWindow();
 });