+++ frontend/app/src/components/accounts/AccountForm.vue
@@ -1,9 +1,9 @@
 <template>
   <v-form ref="form" :value="value" @input="input">
-    <v-row>
+    <v-row no-gutters>
       <v-col cols="12">
         <v-select
-          v-model="selected"
+          v-model="blockchain"
           class="blockchain-balances__chain"
           :items="items"
           :label="$t('account_form.labels.blockchain')"
@@ -16,66 +16,55 @@
             <asset-details class="pt-2 pb-2" :asset="item" />
           </template>
         </v-select>
-        <v-radio-group
-          v-if="isBtc"
-          v-model="btcAccountType"
-          :label="$t('account_form.labels.btc.account_type')"
-          row
-          :disabled="!!edit"
-        >
-          <v-radio value="xpub" :label="$t('account_form.labels.btc.xpub')" />
-          <v-radio
-            value="standalone"
-            :label="$t('account_form.labels.btc.standalone')"
-          />
-        </v-radio-group>
-        <v-row v-if="isBtc && isXpub" align="center" no-gutters>
-          <v-col>
-            <v-text-field
-              v-model="xpub"
-              class="account-form__xpub"
-              :label="$t('account_form.labels.btc.xpub')"
-              autocomplete="off"
-              :disabled="accountOperation || loading || !!edit"
-              @paste="onPasteXpub"
-            />
-          </v-col>
-          <v-col cols="auto">
-            <v-tooltip open-delay="400" top>
-              <template #activator="{ on, attrs }">
-                <v-btn
-                  icon
-                  v-bind="attrs"
-                  v-on="on"
-                  @click="advanced = !advanced"
-                >
-                  <v-icon v-if="advanced">mdi-chevron-up</v-icon>
-                  <v-icon v-else>mdi-chevron-down</v-icon>
-                </v-btn>
-              </template>
-              <span>
-                {{ $tc('account_form.advanced_tooltip', advanced ? 0 : 1) }}
-              </span>
-            </v-tooltip>
-          </v-col>
-        </v-row>
-
-        <v-row v-if="isBtc && btcAccountType === 'xpub' && advanced" no-gutters>
-          <v-col>
-            <v-text-field
-              v-model="derivationPath"
-              class="account-form__derivation-path"
-              :label="$t('account_form.labels.btc.derivation_path')"
-              autocomplete="off"
-              :disabled="accountOperation || loading || !!edit"
-              persistent-hint
-              :hint="$t('account_form.labels.btc.derivation_path_hint')"
-            />
-          </v-col>
-        </v-row>
-
+      </v-col>
+    </v-row>
+    <v-row v-if="!edit" no-gutters>
+      <v-col cols="12">
+        <input-mode-select v-model="inputMode" :blockchain="blockchain" />
+      </v-col>
+    </v-row>
+    <v-row v-if="isBtc && isXpub" align="center" no-gutters class="mt-2">
+      <v-col>
         <v-text-field
-          v-if="!isBtc || ((isBtc && btcAccountType !== 'xpub') || !!edit)"
+          v-model="xpub"
+          class="account-form__xpub"
+          :label="$t('account_form.labels.btc.xpub')"
+          autocomplete="off"
+          :disabled="accountOperation || loading || !!edit"
+          @paste="onPasteXpub"
+        />
+      </v-col>
+      <v-col cols="auto">
+        <v-tooltip open-delay="400" top>
+          <template #activator="{ on, attrs }">
+            <v-btn icon v-bind="attrs" v-on="on" @click="advanced = !advanced">
+              <v-icon v-if="advanced">mdi-chevron-up</v-icon>
+              <v-icon v-else>mdi-chevron-down</v-icon>
+            </v-btn>
+          </template>
+          <span>
+            {{ $tc('account_form.advanced_tooltip', advanced ? 0 : 1) }}
+          </span>
+        </v-tooltip>
+      </v-col>
+    </v-row>
+    <v-row v-if="isBtc && isXpub && advanced" no-gutters>
+      <v-col>
+        <v-text-field
+          v-model="derivationPath"
+          class="account-form__derivation-path"
+          :label="$t('account_form.labels.btc.derivation_path')"
+          autocomplete="off"
+          :disabled="accountOperation || loading || !!edit"
+          persistent-hint
+          :hint="$t('account_form.labels.btc.derivation_path_hint')"
+        />
+      </v-col>
+    </v-row>
+    <v-row no-gutters class="mt-2">
+      <v-col cols="12">
+        <v-text-field
+          v-if="(!isBtc || (isBtc && !isXpub) || !!edit) && !isMetaMask"
           v-model="address"
           class="blockchain-balances__address"
           :label="$t('account_form.labels.account')"
@@ -85,13 +74,25 @@
           :disabled="accountOperation || loading || !!edit"
           @paste="onPasteAddress"
         />
+      </v-col>
+    </v-row>
+    <v-row no-gutters>
+      <v-col cols="12">
         <v-text-field
           v-model="label"
           class="blockchain-balances__label"
           :label="$t('account_form.labels.label')"
           :disabled="accountOperation || loading"
         />
+      </v-col>
+    </v-row>
+    <v-row no-gutters>
+      <v-col cols="12">
         <tag-input v-model="tags" :disabled="accountOperation || loading" />
+      </v-col>
+    </v-row>
+    <v-row no-gutters>
+      <v-col cols="12">
         <div class="blockchain-balances--progress">
           <v-progress-linear v-if="accountOperation" indeterminate />
         </div>
@@ -101,14 +102,25 @@
 </template>
 <script lang="ts">
 import { Component, Emit, Prop, Vue, Watch } from 'vue-property-decorator';
-import { mapGetters } from 'vuex';
+import { mapActions, mapGetters } from 'vuex';
+import {
+  MANUAL_ADD,
+  METAMASK_IMPORT,
+  XPUB_ADD
+} from '@/components/accounts/const';
+import InputModeSelect from '@/components/accounts/InputModeSelect.vue';
+import { AccountInput } from '@/components/accounts/types';
 import TagInput from '@/components/inputs/TagInput.vue';
 import { TaskType } from '@/model/task-type';
 import { deserializeApiErrorMessage } from '@/services/converters';
 import {
   BlockchainAccountPayload,
-  BlockchainAccount
+  BlockchainAccount,
+  AddAccountsPayload,
+  XpubPayload
 } from '@/store/balances/types';
+import { Severity } from '@/store/notifications/consts';
+import { notify } from '@/store/notifications/utils';
 import { Message } from '@/store/types';
 import {
   Blockchain,
@@ -118,20 +130,24 @@ import {
   SupportedBlockchains
 } from '@/typing/types';
 import { trimOnPaste } from '@/utils/event';
+import { getMetamaskAddresses } from '@/utils/metamask';
 
 type ValidationRule = (value: string) => boolean | string;
 
 @Component({
-  components: { TagInput },
+  components: { InputModeSelect, TagInput },
   computed: {
     ...mapGetters('tasks', ['isTaskRunning']),
     ...mapGetters('balances', ['account'])
+  },
+  methods: {
+    ...mapActions('balances', ['addAccounts', 'editAccount', 'addAccount'])
   }
 })
 export default class AccountForm extends Vue {
   readonly items = SupportedBlockchains;
   isTaskRunning!: (type: TaskType) => boolean;
-  selected: Blockchain = ETH;
+  blockchain: Blockchain = ETH;
   pending: boolean = false;
   xpub: string = '';
   derivationPath: string = '';
@@ -140,27 +156,38 @@ export default class AccountForm extends Vue {
   tags: string[] = [];
   errorMessages: string[] = [];
   account!: (address: string) => GeneralAccount | undefined;
+  addAccount!: (payload: BlockchainAccountPayload) => Promise<void>;
+  addAccounts!: (payload: AddAccountsPayload) => Promise<void>;
+  editAccount!: (payload: BlockchainAccountPayload) => Promise<void>;
   advanced: boolean = false;
-
-  btcAccountType: 'xpub' | 'standalone' = 'xpub';
+  inputMode: AccountInput = MANUAL_ADD;
 
   get isBtc(): boolean {
-    return this.selected === BTC;
+    return this.blockchain === BTC;
   }
 
   get isXpub(): boolean {
-    return this.btcAccountType === 'xpub';
+    return this.inputMode === XPUB_ADD;
+  }
+
+  get isMetaMask(): boolean {
+    return this.inputMode === METAMASK_IMPORT;
   }
 
   get rules(): ValidationRule[] {
-    const rules: ValidationRule[] = [
-      (v: string) =>
-        !!v || this.$tc('account_form.validation.address_non_empty')
-    ];
+    const rules: ValidationRule[] = [];
+    if (this.isMetaMask) {
+      return rules;
+    }
+    rules.push(this.nonEmptyRule);
     if (!this.edit) {
       rules.push(this.checkIfExists);
     }
     return rules;
+  }
+
+  private nonEmptyRule(value: string): boolean | string {
+    return !!value || this.$tc('account_form.validation.address_non_empty');
   }
 
   private checkIfExists(value: string): boolean | string {
@@ -182,7 +209,7 @@ export default class AccountForm extends Vue {
     }
 
     this.address = this.edit.address;
-    this.selected = this.edit.chain;
+    this.blockchain = this.edit.chain;
     this.label = this.edit.label;
     this.tags = this.edit.tags;
     if ('xpub' in this.edit) {
@@ -209,6 +236,15 @@ export default class AccountForm extends Vue {
     this.setEditMode();
   }
 
+  @Watch('blockchain')
+  onBlockchainChanged() {
+    if (this.isBtc) {
+      this.inputMode = XPUB_ADD;
+    } else {
+      this.inputMode = MANUAL_ADD;
+    }
+  }
+
   onPasteAddress(event: ClipboardEvent) {
     const paste = trimOnPaste(event);
     if (paste) {
@@ -230,6 +266,8 @@ export default class AccountForm extends Vue {
     this.xpub = '';
     this.derivationPath = '';
     (this.$refs.form as any).resetValidation();
+    this.blockchain = ETH;
+    this.inputMode = MANUAL_ADD;
   }
 
   @Emit()
@@ -250,28 +288,69 @@ export default class AccountForm extends Vue {
     );
   }
 
-  async addAccount(): Promise<boolean> {
-    this.pending = true;
+  async metamaskImport(): Promise<boolean> {
     try {
-      const xpubPayload =
-        this.isBtc && this.btcAccountType === 'xpub'
-          ? {
-              xpub: this.xpub.trim(),
-              derivationPath: this.derivationPath ?? undefined
-            }
-          : undefined;
-      const payload: BlockchainAccountPayload = {
-        blockchain: this.selected,
-        address: this.address.trim(),
-        label: this.label,
-        tags: this.tags,
-        xpub: xpubPayload
-      };
+      let addresses: string[];
+      if (this.$interop.isPackaged) {
+        addresses = await this.$interop.metamaskImport();
+      } else {
+        addresses = await getMetamaskAddresses();
+      }
 
-      await this.$store.dispatch(
-        this.edit ? 'balances/editAccount' : 'balances/addAccount',
-        payload
+      const payload = addresses.map(value => ({
+        address: value,
+        label: this.label,
+        tags: this.tags
+      }));
+
+      await this.addAccounts({
+        blockchain: ETH,
+        payload: payload
+      });
+      return true;
+    } catch (e) {
+      const title = this.$tc('blockchain_balances.metamask_import.error.title');
+      const description = this.$tc(
+        'blockchain_balances.metamask_import.error.description',
+        0,
+        {
+          error: e.message
+        }
       );
+      notify(description, title, Severity.ERROR, true);
+      return false;
+    }
+  }
+
+  payload(): BlockchainAccountPayload {
+    let xpub: XpubPayload | undefined;
+    if (this.isBtc && this.isXpub) {
+      xpub = {
+        xpub: this.xpub.trim(),
+        derivationPath: this.derivationPath ?? undefined
+      };
+    } else {
+      xpub = undefined;
+    }
+
+    return {
+      blockchain: this.blockchain,
+      address: this.address.trim(),
+      label: this.label,
+      tags: this.tags,
+      xpub: xpub
+    };
+  }
+
+  async manualAdd() {
+    const payload = this.payload();
+    try {
+      if (this.edit) {
+        await this.editAccount(payload);
+      } else {
+        await this.addAccount(payload);
+      }
+
       this.reset();
     } catch (e) {
       const apiErrorMessage = deserializeApiErrorMessage(e.message);
@@ -298,8 +377,21 @@ export default class AccountForm extends Vue {
       } as Message);
       return false;
     }
-    this.pending = false;
     return true;
+  }
+
+  async save(): Promise<boolean> {
+    let result: boolean;
+    this.pending = true;
+
+    if (this.isMetaMask) {
+      result = await this.metamaskImport();
+    } else {
+      result = await this.manualAdd();
+    }
+
+    this.pending = false;
+    return result;
   }
 
   private setErrors(errors: string[]) {