+++ frontend/app/src/store/balances/types.d.ts
@@ -54,12 +54,15 @@ interface XpubPayload {
   readonly derivationPath: string;
 }
 
-export interface BlockchainAccountPayload {
-  readonly address: string;
+export interface BlockchainAccountPayload extends AccountPayload {
   readonly blockchain: Blockchain;
+  readonly xpub?: XpubPayload;
+}
+
+export interface AccountPayload {
+  readonly address: string;
   readonly label?: string;
   readonly tags: string[];
-  readonly xpub?: XpubPayload;
 }
 
 export interface ExchangeBalancePayload {
@@ -94,3 +97,8 @@ export interface AssetBalance {
   readonly amount: BigNumber;
   readonly usdValue: BigNumber;
 }
+
+export type AddAccountsPayload = {
+  readonly blockchain: Blockchain;
+  readonly payload: AccountPayload[];
+};